[Organize] Moved Chromatogram smoothing code to one file
